Use  intptr types (simulator) (#20142)

+++ Marlin/src/lcd/dogm/u8g_dev_tft_upscale_from_128x64.cpp
@@ -458,7 +458,7 @@ uint8_t u8g_com_hal_tft_fn(u8g_t *u8g, uint8_t msg, uint8_t arg_val, void *arg_p
     case U8G_COM_MSG_WRITE_SEQ:
       tftio.DataTransferBegin(DATASIZE_16BIT);
       for (uint8_t i = 0; i < arg_val; i += 2)
-        tftio.WriteData(*(uint16_t *)(((uint32_t)arg_ptr) + i));
+        tftio.WriteData(*(uint16_t *)(((uintptr_t)arg_ptr) + i));
       tftio.DataTransferEnd();
       break;
